refactor!: move some function to core/shared
