+++ packages/client/logic/components/index.ts
@@ -1,4 +1,3 @@
 export { ComponentWalker, InstanceMap } from './tree'
 export { getInstanceState, processSetupState, getInstanceDetails, getSetupStateInfo } from './data'
 export { getInstanceOrVnodeRect, getRootElementsFromComponentInstance } from './el'
-export { getInstanceName } from './util'
