+++ packages/core/src/index.ts
@@ -2,3 +2,4 @@ export * from './hook'
 export * from './host'
 export * from './rpc'
 export * from './constant'
+export * from './shared'
