+++ packages/node/src/views/composables.ts
@@ -1,5 +1,6 @@
 import { computed, onMounted, reactive, ref, shallowRef, watchEffect } from 'vue'
 import type { CSSProperties, ComponentInternalInstance, Ref } from 'vue'
+import { getInstanceName } from '@vite-plugin-vue-devtools/core/src'
 import { clamp, createDebounceFn, useObjectStorage, useScreenSafeArea, useWindowEventListener, warn } from './utils'
 
 interface DevToolsFrameState {
@@ -536,7 +537,7 @@ export function useHighlightComponent() {
         const el = instance.vnode.el as HTMLElement | undefined
         if (el && el.nodeType === Node.ELEMENT_NODE) {
           inspectInstance = instance
-          highlight(instance.type.__name ?? '', el.getBoundingClientRect())
+          highlight(getInstanceName(instance), el.getBoundingClientRect())
           return
         }
       }
